rename: index.tsx -> component.tsx 형식으로 변경
