+++ src/widgets/Layout/index.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 
-import { Footer } from './ui/Footer';
-import { Header } from './ui/Header';
+import { Footer } from './ui/Footer/Footer';
+import { Header } from './ui/Header/Header';
 
 export const Layout = () => {
   return (
