+++ src/widgets/Layout/ui/Footer/Footer.module.scss
