+++ src/widgets/Layout/ui/Footer/Footer.tsx
@@ -1,4 +1,4 @@
-import styles from './index.module.scss';
+import styles from './Footer.module.scss';
 
 export const Footer = () => {
   return (
