+++ src/widgets/Layout/ui/Header/Header.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
 
-import styles from './index.module.scss';
+import styles from './Header.module.scss';
 import { NAV_LINKS } from '../../constants';
 
 //assets
